Fix Beaver Builder conflict.
Previously, the trestle_tiny_mce_before_init filter was causing a fatal
error when it was trying to call get_current_screen() on the front-end,
which isn't availably unless /wp-admin/includes/screen.php is included,
which it isn't on the front end. This patch switches from checking the
current screen to simply checking if the post object exists. Not sure if
there are any edge cases where the current screen check does something
the post check doesn't - I'll leave that to you @braad to decide as I
think you added this code.